Move module twice_differentiable to top level of influence package
