+++ src/pydvl/influence/twice_differentiable.py
